ai_platform/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# AI & NLP
nltk==3.8.1
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# CLI
click==8.1.7
requests==2.31.0

# Voice processing
speechrecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.11

# 3D Visualization
pygame==2.5.1
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0